Updates to support round-trip between orig. SD 2.0 and diffusers models
