+++ scripts/convert_original_stable_diffusion_to_diffusers.py
@@ -647,8 +647,72 @@ def convert_ldm_clip_checkpoint(checkpoint):
     return text_model
 
 
+
+
+import re
+
+textenc_conversion_lst = [
+    ('cond_stage_model.model.positional_embedding',
+                "text_model.embeddings.position_embedding.weight"),
+    ('cond_stage_model.model.token_embedding.weight',
+                "text_model.embeddings.token_embedding.weight"),
+    ('cond_stage_model.model.ln_final.weight', 'text_model.final_layer_norm.weight'),
+    ('cond_stage_model.model.ln_final.bias', 'text_model.final_layer_norm.bias')
+]
+textenc_conversion_map = {x[0]:x[1] for x in textenc_conversion_lst}
+
+textenc_transformer_conversion_lst = [
+    # (stable-diffusion, HF Diffusers)
+    ('resblocks.','text_model.encoder.layers.'),
+    ('ln_1','layer_norm1'),
+    ('ln_2','layer_norm2'),
+    ('.c_fc.','.fc1.'),
+    ('.c_proj.','.fc2.'),
+    ('.attn','.self_attn'),
+    ('ln_final.','transformer.text_model.final_layer_norm.'),
+    ('token_embedding.weight','transformer.text_model.embeddings.token_embedding.weight'),
+    ('positional_embedding','transformer.text_model.embeddings.position_embedding.weight')
+]
+protected = {re.escape(x[0]):x[1] for x in textenc_transformer_conversion_lst}
+textenc_pattern = re.compile("|".join(protected.keys()))
+
 def convert_open_clip_checkpoint(checkpoint):
     text_model = CLIPTextModel.from_pretrained("stabilityai/stable-diffusion-2", subfolder="text_encoder")
+
+    keys = list(checkpoint.keys())
+
+    text_model_dict = {}
+
+    d_model = int( checkpoint['cond_stage_model.model.text_projection'].shape[0] )
+
+    text_model_dict["text_model.embeddings.position_ids"] = \
+                    text_model.text_model.embeddings.get_buffer('position_ids')
+
+    for key in keys:
+        if 'resblocks.23' in key:   # Diffusers drops the final layer and only uses the penultimate layer
+            continue
+        if key in textenc_conversion_map:
+            text_model_dict[textenc_conversion_map[key]] = checkpoint[key]
+        if key.startswith("cond_stage_model.model.transformer."):
+            new_key = key[len("cond_stage_model.model.transformer.") :]
+            if new_key.endswith(".in_proj_weight"):
+                new_key = new_key[:-len(".in_proj_weight")]
+                new_key = textenc_pattern.sub(lambda m: protected[re.escape(m.group(0))], new_key)
+                text_model_dict[new_key+'.q_proj.weight'] = checkpoint[key][:d_model,:]
+                text_model_dict[new_key+'.k_proj.weight'] = checkpoint[key][d_model:d_model*2,:]
+                text_model_dict[new_key+'.v_proj.weight'] = checkpoint[key][d_model*2:,:]
+            elif new_key.endswith(".in_proj_bias"):
+                new_key = new_key[:-len(".in_proj_bias")]
+                new_key = textenc_pattern.sub(lambda m: protected[re.escape(m.group(0))], new_key)
+                text_model_dict[new_key+'.q_proj.bias'] = checkpoint[key][:d_model]
+                text_model_dict[new_key+'.k_proj.bias'] = checkpoint[key][d_model:d_model*2]
+                text_model_dict[new_key+'.v_proj.bias'] = checkpoint[key][d_model*2:]
+            else:
+                new_key = textenc_pattern.sub(lambda m: protected[re.escape(m.group(0))], new_key)
+
+                text_model_dict[new_key] = checkpoint[key]
+
+    text_model.load_state_dict(text_model_dict)
 
     # SKIP for now - need openclip -> HF conversion script here
     #    keys = list(checkpoint.keys())
@@ -678,7 +742,7 @@ if __name__ == "__main__":
     )
     parser.add_argument(
         "--scheduler_type",
-        default="pndm",
+        default="default",    # will be either PNDM (for SD 1.x) or DDPM (for SD 2.0)
         type=str,
         help="Type of scheduler to use. Should be one of ['pndm', 'lms', 'ddim', 'euler', 'euler-ancest', 'dpm']",
     )
@@ -769,6 +833,15 @@ if __name__ == "__main__":
         set_alpha_to_one=False,
         prediction_type=prediction_type,
     )
+
+    if args.scheduler_type == "default":
+# All of the 2.0 models use OpenCLIP and all use DDPM scheduler by default.
+        text_model_type = original_config.model.params.cond_stage_config.target.split(".")[-1]
+        if text_model_type == "FrozenOpenCLIPEmbedder":
+            args.scheduler_type = "ddim"
+        else:
+            args.scheduler_type = "pndm"
+
     if args.scheduler_type == "pndm":
         config = dict(scheduler.config)
         config["skip_prk_steps"] = True
